feat: database changes for multi account
