--- src/constants/BtcCoinData.ts
@@ -1,7 +1,0 @@
-import { CoinData, ICoinDataOptions } from './CoinData';
-
-export class BtcCoinData extends CoinData {
-  constructor(coinData: ICoinDataOptions) {
-    super({ ...coinData });
-  }
-}
