+++ src/constants/types/AbsCoinData.ts
