+++ src/constants/types/Erc20CoinData.ts
