+++ src/constants/types/EthCoinData.ts
