+++ src/constants/types/NearCoinData.ts
