+++ src/constants/types/SolanaCoinData.ts
